requirements.txt
Flask
flask_cors
Gunicorn
python-dotenv
requests
